RDIStitcher/requirements.txt
absl-py==2.1.0
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albumentations==1.3.0
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
basicsr==1.4.2
bitsandbytes==0.43.3
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
cityscapesScripts==2.2.1
click==8.1.7
clip
cmake==3.30.3
CoCa-pytorch==0.1.0
coloredlogs==15.0.1
cycler==0.12.1
diffusers==0.20.1
distro==1.9.0
einops==0.7.0
exceptiongroup==1.2.1
fastapi==0.114.0
ffmpy==0.4.0
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.1.1
future==1.0.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
gradio==3.44.2
gradio_client==0.5.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.35.1
importlib_metadata==7.1.0
importlib_resources==6.4.5
jieba==0.42.1
Jinja2==3.1.3
joblib==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kornia==0.6.2
lit==18.1.8
lmdb==1.3.0
loguru==0.6.0
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.3.4
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.6.4
networkx==3.3
numpy==1.22.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.12.0
open-clip-torch==2.16.0
openai==1.30.5
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.64
orjson==3.10.7
packaging==20.9
pandas==2.0.3
parrots==0.1.7
peft==0.5.0
Pillow==9.5.0
prettytable==3.2.0
protobuf==3.20.1
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.2
pydantic_core==2.18.3
pyDeprecate==0.3.1
pydub==0.25.1
pyparsing==3.1.4
pypinyin==0.52.0
pyquaternion==0.9.9
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorch-lightning==1.5.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
referencing==0.35.1
regex==2021.4.4
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.3.1
scikit-image==0.18.1
scikit-learn==1.1.3
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.12.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.5
test-tube==0.7.5
threadpoolctl==3.5.0
tifffile==2020.10.1
timm==0.4.12
tokenizers==0.15.2
torch==2.0.0
torch-tb-profiler==0.4.1
torchmetrics==0.7.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.36.0
triton==2.0.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.6
vit-pytorch==1.7.12
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.4
xformers==0.0.17
yacs==0.1.8
yapf==0.31.0
yarl==1.11.1
zipp==3.19.0
